requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
python-docx==1.1.0
PyPDF2==3.0.1
pdf2docx==0.5.6
docx2pdf==0.1.8
pymongo==4.6.0
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1
PyMuPDF==1.23.0
qrcode==7.4.2
gtts==2.4.0
pygame==2.5.2
aiofiles==23.2.1
pytest==7.4.0
pytest-cov==4.1.0
